# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Way Bill Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  code: string;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  gstin?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  isActive: boolean;
  createdAt: string;
}