# lib/api-zod/src/generated/types/companyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Way Bill Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanyUpdate {
  name?: string;
  address?: string;
  gstin?: string;
  contactEmail?: string;
  contactPhone?: string;
  isActive?: boolean;
}