# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Way Bill Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { EwayBill } from './ewayBill';

export interface DashboardStats {
  totalCompanies: number;
  totalUsers: number;
  totalEwaybills: number;
  recentEwaybills: EwayBill[];
}