# lib/api-zod/src/generated/types/ewayBill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Way Bill Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface EwayBill {
  id: number;
  ewbNumber: string;
  companyId: number;
  userId: number;
  generatedBy?: string;
  generatedDate: string;
  validUpto: string;
  supplyType: string;
  transactionType: string;
  fromGstin: string;
  fromTradeName: string;
  fromAddr1: string;
  /** @nullable */
  fromAddr2?: string | null;
  fromPincode: string;
  fromStateCode: string;
  toGstin: string;
  toTradeName: string;
  toAddr1: string;
  /** @nullable */
  toAddr2?: string | null;
  toPincode: string;
  toStateCode: string;
  itemName: string;
  hsnCode: string;
  quantity: number;
  unit: string;
  taxableValue: number;
  cgstRate: number;
  sgstRate: number;
  igstRate: number;
  /** @nullable */
  cessRate?: number | null;
  totalValue: number;
  transporterDocNo: string;
  transporterDocDate: string;
  vehicleNo: string;
  vehicleType: string;
  transportMode: string;
  /** @nullable */
  distance?: number | null;
  /** @nullable */
  transactionSubType?: string | null;
  status: string;
  createdAt: string;
}